refactor(lib): DB・キーストローク監視処理を別モジュールに分割
- src/db.rs に DB 操作（init_db / flush_minute_count / query_today_db_count）を集約
- src/keystroke.rs にキーストローク監視（macOS: CGEventTap / その他: rdev）を集約
- src/lib.rs はアプリライフサイクル（setup・RunEvent ハンドリング）に特化し 110 行に削減
